# openapi.yaml
openapi: 3.0.1
info:
  title: Magic Plugin
  description: |
    API for retrieving Kaltura media entries and their captions.
    This API allows you to search for media entries and fetch captions for specific entries.
  version: 1.0.0
servers:
  - url: https://copilot-server-test.onrender.com

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    KalturaEntry:
      type: object
      properties:
        id:
          type: string
          description: Kaltura entry ID
          example: "1_abc123xyz"
        name:
          type: string
          description: Entry title
          example: "Alzheimer's Disease Overview"
        description:
          type: string
          description: Entry description
          example: "A comprehensive overview of Alzheimer's disease"
        thumbnailUrl:
          type: string
          description: URL to the entry thumbnail
        createdAt:
          type: integer
          description: Creation timestamp
        updatedAt:
          type: integer
          description: Last update timestamp
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Check if server is running
      operationId: healthCheck
      responses:
        '200':
          description: Server is operational
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  /kaltura/entries:
    get:
      summary: Fetch Kaltura entries using KS from token
      description: Extracts the Kaltura Session (KS) from the JWT token and uses it to fetch entries from Kaltura
      operationId: getKalturaEntries
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved Kaltura entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  objects:
                    type: array
                    items:
                      $ref: '#/components/schemas/KalturaEntry'
        '400':
          description: Bad request - No KS found in token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error when calling Kaltura API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /magic-response:
    post:
      summary: Generate a magical answer to any question
      operationId: getMagicAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                  description: The user's question
              required:
                - question
      responses:
        '200':
          description: The magic answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    example: "🎩 Hello from Amir's Plugin! You asked: 'X'"

  /kaltura/captions/{entryId}:
    get:
      summary: Fetch captions for a specific Kaltura entry
      description: |
        Gets the caption content for a specific entry ID.
        This endpoint handles the complete process:
        1. Fetches the caption assets for the entry
        2. Gets the download URL for the caption file
        3. Downloads and returns the actual caption content
      operationId: getKalturaCaptions
      security:
        - BearerAuth: []
      parameters:
        - name: entryId
          in: path
          required: true
          description: The ID of the Kaltura entry to fetch captions for
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved captions content
          content:
            application/json:
              schema:
                type: object
                properties:
                  caption_id:
                    type: string
                    description: ID of the caption asset
                    example: "0_captionid123"
                  entry_id:
                    type: string
                    description: ID of the entry
                    example: "0_entryid456"
                  format:
                    type: string
                    description: Format of the caption
                    example: "srt"
                  language:
                    type: string
                    description: Language of the caption
                    example: "en"
                  content:
                    type: string
                    description: The actual caption text content
                    example: "1\n00:00:00,000 --> 00:00:05,000\nWelcome to this video about Alzheimer's disease."
        '400':
          description: Bad request - No KS found in token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No captions found for this entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error when calling Kaltura API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'